+++ src/main/java/org/percepta/mgrankvi/floorplanner/gwt/client/item/table/CTable.java
@@ -63,7 +63,7 @@ public class CTable extends VisualItem {
 	}
 
 	@Override
-	public boolean pointInObject(final int x, final int y) {
+	public boolean pointInObject(final double x, final double y) {
 		if (points.isEmpty()) {
 			return false;
 		}
